Mithun dev

+++ superSenseTag.py
@@ -2,6 +2,7 @@ from tqdm import tqdm
 import json,mmap,os,argparse
 import processors
 from processors import *
+from nltk.tokenize import sent_tokenize
 
 def parse_commandline_args():
     return create_parser().parse_args()
@@ -17,6 +18,10 @@ def create_parser():
                         help='use docker for loading pyproc. useful in machines where you have root access.', metavar='BOOL')
     parser.add_argument('--output_folder', type=str, default='outputs/',
                         help='folder where outputs will be created')
+    parser.add_argument('--write_pos_tags', default=False, type=str2bool,
+                        help='if you want to write the pos tags to disk or not. This is used as input for the SSTagger.',
+                        metavar='BOOL')
+
 
     print(parser.parse_args())
     return parser
@@ -58,6 +63,10 @@ def read_rte_data(filename):
 
     return all_claims, all_evidences, all_labels
 
+
+
+
+
 def annotate(headline, body, API):
     claim_ann = API.fastnlp.annotate(headline)
     ev_ann = API.fastnlp.annotate(body)
@@ -92,13 +101,23 @@ def write_json_to_disk(claim, evidence,label,outfile):
     outfile.write('\n')
 
 def write_token_POS_disk_as_csv(annotated_sent,full_path_output_file):
-    # write one file per claim. empty the file out if it already exists.
-    with open(full_path_output_file, 'w') as outfile:
-        outfile.write('')
-        with open(full_path_output_file, 'a+') as outfile:
-            for word,postag in zip(annotated_sent.words, annotated_sent.tags):
-                outfile.write(word+"\t"+postag+"\n")
+    # if the file already exists, leave it. It might have been written in a run before
+    if not (os.path.isfile(full_path_output_file)):
+        with open(full_path_output_file, 'w') as outfile:
+            outfile.write('')
+            with open(full_path_output_file, 'a+') as outfile:
+                for word,postag in zip(annotated_sent.words, annotated_sent.tags):
+                    outfile.write(word+"\t"+postag+"\n")
 
+def write_list_of_evidence_sentences_to_disk(annotated_sentences,full_path_output_file):
+        if not (os.path.isfile(full_path_output_file)):
+            with open(full_path_output_file, 'w') as outfile:
+                outfile.write('')
+                with open(full_path_output_file, 'a+') as outfile:
+                    for annotated_sent in annotated_sentences:
+                        for word, postag in zip(annotated_sent.words, annotated_sent.tags):
+                            outfile.write(word + "\t" + postag + "\n")
+                        outfile.write("\n")
 
 if __name__ == '__main__':
 
@@ -108,7 +127,8 @@ if __name__ == '__main__':
     else:
         API = ProcessorsAPI(port=args.pyproc_port)
 
-    filename=args.inputFile
+    cwd=os.getcwd()
+    filename=cwd+"/"+args.inputFile
     all_claims, all_evidences, all_labels=read_rte_data(filename)
     all_claims_neutered=[]
 
@@ -116,19 +136,32 @@ if __name__ == '__main__':
 
 
 
-    for (index, (c, e ,l)) in enumerate(zip(all_claims, all_evidences,all_labels)):
+    length=len(all_evidences)
+    index=0
 
-            claim_ann, ev_ann = annotate(c, e, API)
-            assert (claim_ann is not None)
+    for (c, e ,l) in tqdm(zip(all_claims, all_evidences,all_labels),total=length):
+        claim_ann = API.fastnlp.annotate(c)
+        assert (claim_ann is not None)
+
+        # new requirement. SSTagger wants a newline after every sentence. So will have to find sentence boundaries.
+        evidence_sentences=sent_tokenize(e)
+        evidence_sentences_annotated=[]
+        for evidence_sentence in evidence_sentences:
+            ev_ann = API.fastnlp.annotate(evidence_sentence)
             assert (ev_ann is not None)
+            evidence_sentences_annotated.append(ev_ann)
 
-            # write each token and its pos tag to disk, with one line each
+        if(args.write_pos_tags):
+            # write each token and its pos tag to disk, with one line each-This is used as input for the SSTagger
             out_file_name="claim_words_pos_datapointid_"+str(index)
             full_path_output_file=output_folder+out_file_name
             write_token_POS_disk_as_csv(claim_ann, full_path_output_file)
+
             out_file_name = "evidence_words_pos_datapointid_" + str(index)
             full_path_output_file = output_folder + out_file_name
-            write_token_POS_disk_as_csv(ev_ann, full_path_output_file)
+            write_list_of_evidence_sentences_to_disk(evidence_sentences_annotated, full_path_output_file)
+            index=index+1
+
